Moved sequel-waterline-orientdb initialisation and reference to connection Each collection no longer holds a reference to waterline.schema
