+++ lib/adapter.js
@@ -4,7 +4,8 @@
  */
 var Connection = require('./connection'),
     Associations = require('./associations'),
-    utils = require('./utils');
+    utils = require('./utils'),
+    log = require('debug-logger')('waterline-orientdb:adapter');
 
 /**
  * waterline-orientdb
@@ -91,8 +92,6 @@ module.exports = (function() {
         // Waterline only allows populating 1 level below. fetchPlanLevel allows to
         // to populate further levels below
         fetchPlanLevel : 1,
-        // Turns on parameterized queries and it should be enabled, but breaks 2 tests against 1.7.*.
-        // https://github.com/appscot/waterline-orientdb/issues/20
         parameterized : true
       }
     },
@@ -109,6 +108,7 @@ module.exports = (function() {
      * @return {[type]} [description]
      */
     registerConnection : function(connection, collections, cb) {
+      log.debug('registerConnection:', connection.database);
 
       if (!connection.identity)
         return cb(new Error('Connection is missing an identity.'));
@@ -163,6 +163,7 @@ module.exports = (function() {
      * @param {Function} callback
      */
     describe : function(connection, collection, cb) {
+      log.debug('describe:', collection);
       // Add in logic here to describe a collection (e.g. DESCRIBE TABLE
       // logic)
       connections[connection].describe(collection, cb);
@@ -182,6 +183,7 @@ module.exports = (function() {
      * @param {Function} cb
      */
     define : function(connection, collection, definition, cb) {
+      log.debug('define:', collection);
 
       // Create the collection and indexes
       connections[connection].createCollection(collection, definition, cb);
@@ -199,6 +201,7 @@ module.exports = (function() {
      * @param {Function} callback
      */
     drop : function(connection, collection, relations, cb) {
+      log.debug('drop:', collection);
       // Add in logic here to delete a collection (e.g. DROP TABLE logic)
 
       return connections[connection].drop(collection, relations, cb);
