+++ lib/collection/document.js
@@ -38,9 +38,6 @@ var Document = module.exports = function Document(definition, connection, databa
   
   // Hold Schema in OrientDB friendly format
   this.orientdbSchema = null;
-  
-  // Hold the waterline schema, used by query namely waterline-sequel-orientdb
-  this.waterlineSchema = null;
 
   // Hold a reference to an active connection
   this.connection = connection;
@@ -76,7 +73,7 @@ Document.prototype.find = function find(criteria, cb) {
   var _query, query;
   
   try {
-    _query = new Query(criteria, self.waterlineSchema, self.connection.config);
+    _query = new Query(criteria, self.connection.sequel);
   } catch(err) { return cb(err); }
   
   try {
@@ -156,7 +153,7 @@ Document.prototype.update = function update(criteria, values, cb) {
 
   // Catch errors from building query and return to the callback
   try {
-    _query = new Query(criteria, self.waterlineSchema, self.connection.config);
+    _query = new Query(criteria, self.connection.sequel);
     _document = new Doc(values, self.schema, self.connection);
     where = _query.getWhereQuery(self.tableName);
   } catch(e) {
@@ -251,10 +248,8 @@ Document.prototype._parseDefinition = function _parseDefinition(definition, coll
   // Set the identity
   var identity = definition.identity ? definition.identity : definition.tableName;
   this.identity = _.clone(identity);
-  
-  this.waterlineSchema = definition.waterline.schema;
-  
-  
+
+  // Create orientdbSchema
   this.orientdbSchema = {};
   
   Object.keys(self.schema).forEach(function(attributeName) {
@@ -340,7 +335,6 @@ Document.prototype._parseDefinition = function _parseDefinition(definition, coll
  *
  * @api private
  */
-
 Document.prototype._buildIndexes = function _buildIndexes() {
   var self = this;
