+++ lib/connection.js
@@ -7,8 +7,18 @@ var Oriento = require('oriento'),
     utils = require('./utils'),
     Associations = require('./associations'),
     log = require('debug-logger')('waterline-orientdb:connection'),
-    Collection = require('./collection');
+    Collection = require('./collection'),
+    Sequel = require('waterline-sequel-orientdb');
 
+// waterline-sequel-orientdb options
+var sqlOptions = {
+  parameterized : true,
+  caseSensitive : false,
+  escapeCharacter : '',
+  casting : false,
+  canReturnValues : true,
+  escapeInserts : true
+}; 
 
 module.exports = (function () {
 
@@ -38,6 +48,16 @@ module.exports = (function () {
         this.associations = new Associations(this);
         this.server = server;
         
+        // update sqlOptions config
+        sqlOptions.parameterized = _.isUndefined(this.config.options.parameterized) ?
+                                   sqlOptions.parameterized : this.config.options.parameterized;
+        // Hold the waterline schema, used by query namely waterline-sequel-orientdb
+        this.waterlineSchema = _.values(collections)[0].waterline.schema;
+        
+        // Instantiate a sequel helper
+        this.sequel = new Sequel(this.waterlineSchema, sqlOptions);
+        
+        // Stores existings classes from OrientDB
         this.dbClasses = {};
         classes.forEach(function(klass){
           self.dbClasses[klass.name] = klass;
