+++ lib/document.js
@@ -34,7 +34,7 @@ var Document = module.exports = function Document(values, schema, connection) {
   if(values){
     var newValues = this.setValues(values);
     this.values = newValues.values;
-  } 
+  }
 
   return this;
 };
