+++ lib/query.js
@@ -5,19 +5,7 @@
 var _ = require('lodash'),
     utils = require('./utils'),
     hop = utils.object.hop,
-    Sequel = require('waterline-sequel-orientdb'),
     RID = require('oriento').RID;
-    
-
-// waterline-sequel-orientdb options
-var sqlOptions = {
-  parameterized : true,
-  caseSensitive : false,
-  escapeCharacter : '',
-  casting : false,
-  canReturnValues : true,
-  escapeInserts : true
-}; 
 
 
 /**
@@ -28,19 +16,13 @@ var sqlOptions = {
  * @param {Object} options
  * @api private
  */
-var Query = module.exports = function Query(options, schema, adapterConfig) {
-  // Apply configs
-  sqlOptions.parameterized = _.isUndefined(adapterConfig.options.parameterized) ?
-                             sqlOptions.parameterized : adapterConfig.options.parameterized;
-
-  // Cache the schema for use in parseTypes
-  this.schema = schema;
-
+var Query = module.exports = function Query(options, sequel) {
+  
+  // Sequel builder
+  this.sequel = sequel;
+  
   // Normalize Criteria
   this.criteria = this.normalizeCriteria(options);
-  
-  // Instantiate a sequel helper
-  this.sequel = new Sequel(schema, sqlOptions);
 
   return this;
 };
